Moved Controller annotation to javax.mvc package
