+++ api/src/main/java/javax/mvc/Controller.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package javax.mvc.annotation;
+package javax.mvc;
 
 import javax.ws.rs.NameBinding;
 import java.lang.annotation.Documented;
